pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "surfacia"
version = "3.0.1"
authors = [
    {name = "YumingSu", email = "823808458@qq.com"},
]
description = "A comprehensive Python package for molecular structure-activity relationship (QSAR/QSPR) studies with interactive SHAP visualization and AI-powered analysis"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
keywords = [
    "chemistry",
    "quantum-chemistry", 
    "molecular-descriptors",
    "QSAR",
    "QSPR",
    "computational-chemistry",
    "rdkit",
    "gaussian",
    "multiwfn",
    "machine-learning",
    "SHAP",
    "molecular-visualization"
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "openbabel-wheel>=3.1.1",
    "scikit-learn>=1.0.0",
    "xgboost>=1.6.0",
    "shap>=0.41.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "py3Dmol>=2.0.0",
    "ipywidgets>=7.6.0",
    "plotly>=5.0.0",
    "zhipuai>=2.0.0",
    "dash>=2.0.0",
    "cclib>=1.7.2",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "toml>=0.10.2",
    "tqdm>=4.64.0",
    "pathlib2>=2.3.7",
    "joblib>=1.1.0",
    "jupyter>=1.0.0",
    "notebook>=6.4.0",
    "ipython>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
]
viz = [
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
]

[project.scripts]
surfacia = "surfacia.cli:main"

[project.urls]
Homepage = "https://github.com/sym823808458/Surfacia"
"Bug Reports" = "https://github.com/sym823808458/Surfacia/issues"
"Source" = "https://github.com/sym823808458/Surfacia"
"Documentation" = "https://surfacia.readthedocs.io/"

[tool.setuptools]
packages = ["surfacia"]
include-package-data = true

[tool.setuptools.package-data]
surfacia = [
    "data/*.csv",
    "data/*.json", 
    "config/*.yaml",
    "templates/*.txt",
]

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
